EBAC/Projeto_Final/requirements.txt
pandas==1.3.3
numpy==1.21.2
numpydoc==1.5.0
pandocfilters==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
python-snappy==0.6.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
plotly==5.9.0
streamlit==1.28.1
streamlit-pandas-profiling==0.1.3
Pillow==8.4.0
scikit-learn==0.24.2
